Smaller test names to not trip MSMQ
