+++ src/ServiceControl.AcceptanceTests/HeartbeatMonitoring/When_endpoints_without_heartbeats_starts_up.cs
@@ -4,24 +4,23 @@
     using System.Linq;
     using System.Threading.Tasks;
     using EndpointTemplates;
-    using Infrastructure.Settings;
     using NServiceBus;
     using NServiceBus.AcceptanceTesting;
     using NUnit.Framework;
     using ServiceControl.CompositeViews.Endpoints;
     using Conventions = NServiceBus.AcceptanceTesting.Customization.Conventions;
 
-    class When_an_unmonitored_endpoint_starts_sending_heartbeats : AcceptanceTest
+    class When_endpoints_without_heartbeats_starts_up : AcceptanceTest
     {
-        static string EndpointName => Conventions.EndpointNamingConvention(typeof(MyEndpoint));
+        static string EndpointName => Conventions.EndpointNamingConvention(typeof(StartingEndpoint));
 
         [Test]
-        public async Task Should_be_marked_as_monitored()
+        public async Task Should_not_be_monitored()
         {
             List<EndpointsView> endpoints = null;
 
             await Define<MyContext>()
-                .WithEndpoint<MyEndpoint>(c => c.When(bus => bus.SendLocal(new MyMessage())))
+                .WithEndpoint<StartingEndpoint>(c => c.When(bus => bus.SendLocal(new MyMessage())))
                 .Done(async c =>
                 {
                     var result = await this.TryGetMany<EndpointsView>("/api/endpoints/", e => e.Name == EndpointName);
@@ -33,30 +32,15 @@
             var myEndpoint = endpoints.FirstOrDefault(e => e.Name == EndpointName);
             Assert.NotNull(myEndpoint);
             Assert.IsFalse(myEndpoint.Monitored);
-
-            await Define<MyContext>()
-                .WithEndpoint<MyEndpointWithHeartbeat>()
-                .Done(async c =>
-                {
-                    var result = await this.TryGetMany<EndpointsView>("/api/endpoints/", e => e.Name == EndpointName && e.Monitored && e.MonitorHeartbeat && e.IsSendingHeartbeats);
-                    endpoints = result;
-                    return result;
-                })
-                .Run();
-
-            myEndpoint = endpoints.SingleOrDefault(e => e.Name == EndpointName);
-            Assert.NotNull(myEndpoint);
-            Assert.IsTrue(myEndpoint.Monitored);
-            Assert.IsTrue(myEndpoint.IsSendingHeartbeats);
         }
 
         public class MyContext : ScenarioContext
         {
         }
 
-        public class MyEndpoint : EndpointConfigurationBuilder
+        public class StartingEndpoint : EndpointConfigurationBuilder
         {
-            public MyEndpoint()
+            public StartingEndpoint()
             {
                 EndpointSetup<DefaultServerWithAudit>();
             }
@@ -67,14 +51,6 @@
                 {
                     return Task.FromResult(0);
                 }
-            }
-        }
-
-        public class MyEndpointWithHeartbeat : EndpointConfigurationBuilder
-        {
-            public MyEndpointWithHeartbeat()
-            {
-                EndpointSetup<DefaultServerWithoutAudit>(c => { c.SendHeartbeatTo(Settings.DEFAULT_SERVICE_NAME); }).CustomEndpointName(EndpointName);
             }
         }
