+++ src/ServiceControl.AcceptanceTests/MultiInstance/When_message_searched_by_conversationId.cs
@@ -1,8 +1,6 @@
 ﻿namespace ServiceBus.Management.AcceptanceTests.MultiInstance
 {
     using System.Collections.Generic;
-    using System.Linq;
-    using System.Net.Http;
     using System.Threading.Tasks;
     using EndpointTemplates;
     using Infrastructure.Settings;
@@ -10,43 +8,26 @@
     using NServiceBus.AcceptanceTesting;
     using NServiceBus.AcceptanceTesting.Customization;
     using NServiceBus.AcceptanceTests;
-    using NServiceBus.Settings;
     using NUnit.Framework;
-    using ServiceControl.CompositeViews.Endpoints;
+    using ServiceControl.CompositeViews.Messages;
 
-    class When_processed_message_multi_instance_endpoint_processing : AcceptanceTest
+    class When_message_searched_by_conversationId : AcceptanceTest
     {
         [Test]
-        public async Task Should_be_listed_in_known_endpoints()
+        public async Task Should_be_found()
         {
             SetInstanceSettings = ConfigureRemoteInstanceForMasterAsWellAsAuditAndErrorQueues;
 
-            List<KnownEndpointsView> knownEndpoints = null;
-            HttpResponseMessage httpResponseMessage = null;
-
-            var context = await Define<MyContext>(Remote1, Master)
-                .WithEndpoint<Sender>(b => b.When((bus, c) => bus.Send(new MyMessage())))
+            await Define<MyContext>(Remote1, Master)
+                .WithEndpoint<Sender>(b => b.When((bus, c) => bus.SendLocal(new TriggeringMessage())))
                 .WithEndpoint<ReceiverRemote>()
                 .Done(async c =>
                 {
-                    var result = await this.TryGetMany<KnownEndpointsView>("/api/endpoints/known", m => m.EndpointDetails.Name == c.EndpointNameOfReceivingEndpoint, Master);
-                    knownEndpoints = result;
-                    if (result)
-                    {
-                        httpResponseMessage = await this.GetRaw("/api/endpoints/known", Master);
-
-                        return true;
-                    }
-
-                    return false;
+                    var result = await this.TryGetMany<MessagesView>($"/api/conversations/{c.ConversationId}", instanceName: Master);
+                    List<MessagesView> response = result;
+                    return c.ConversationId != null && result && response.Count == 2;
                 })
                 .Run();
-
-            Assert.AreEqual(context.EndpointNameOfReceivingEndpoint, knownEndpoints.Single(e => e.EndpointDetails.Name == context.EndpointNameOfReceivingEndpoint).EndpointDetails.Name);
-            Assert.AreEqual(ReceiverHostDisplayName, knownEndpoints.Single(e => e.EndpointDetails.Name == context.EndpointNameOfReceivingEndpoint).HostDisplayName);
-
-            Assert.NotNull(httpResponseMessage);
-            Assert.False(httpResponseMessage.Headers.Contains("ETag"), "Expected not to contain ETag header, but it was found.");
         }
 
         private void ConfigureRemoteInstanceForMasterAsWellAsAuditAndErrorQueues(string instanceName, Settings settings)
@@ -76,7 +57,6 @@
         private string addressOfRemote;
         private const string Master = "master";
         private const string Remote1 = "remote1";
-        private const string ReceiverHostDisplayName = "Rico";
         private static string AuditMaster = $"{Master}.audit";
         private static string ErrorMaster = $"{Master}.error";
         private static string AuditRemote = $"{Remote1}.audit";
@@ -93,21 +73,18 @@
 
                     c.ConfigureTransport()
                         .Routing()
-                        .RouteToEndpoint(typeof(MyMessage), typeof(ReceiverRemote));
+                        .RouteToEndpoint(typeof(TriggeredMessage), typeof(ReceiverRemote));
                 });
             }
 
-            public class MyMessageHandler : IHandleMessages<MyMessage>
+            public class TriggeringMessageHandler : IHandleMessages<TriggeringMessage>
             {
                 public MyContext Context { get; set; }
 
-                public ReadOnlySettings Settings { get; set; }
-
-                public Task Handle(MyMessage message, IMessageHandlerContext context)
+                public Task Handle(TriggeringMessage message, IMessageHandlerContext context)
                 {
-                    Context.EndpointNameOfReceivingEndpoint = Settings.EndpointName();
-                    Context.MasterMessageId = context.MessageId;
-                    return Task.FromResult(0);
+                    Context.ConversationId = context.MessageHeaders[Headers.ConversationId];
+                    return context.Send(new TriggeredMessage());
                 }
             }
         }
@@ -120,38 +97,31 @@
                 {
                     c.AuditProcessedMessagesTo(AuditRemote);
                     c.SendFailedMessagesTo(ErrorRemote);
-
-                    c.UniquelyIdentifyRunningInstance().UsingCustomDisplayName(ReceiverHostDisplayName);
                 });
             }
 
-            public class MyMessageHandler : IHandleMessages<MyMessage>
+            public class TriggeredMessageHandler : IHandleMessages<TriggeredMessage>
             {
                 public MyContext Context { get; set; }
 
-                public ReadOnlySettings Settings { get; set; }
-
-                public Task Handle(MyMessage message, IMessageHandlerContext context)
+                public Task Handle(TriggeredMessage message, IMessageHandlerContext context)
                 {
-                    Context.EndpointNameOfReceivingEndpoint = Settings.EndpointName();
-                    Context.Remote1MessageId = context.MessageId;
+                    Context.ConversationId = context.MessageHeaders[Headers.ConversationId];
                     return Task.FromResult(0);
                 }
             }
         }
 
+        public class TriggeringMessage : ICommand
+        {
+        }
 
-        public class MyMessage : ICommand
+        public class TriggeredMessage : ICommand
         {
         }
 
         public class MyContext : ScenarioContext
         {
-            public string MasterMessageId { get; set; }
-            public string Remote1MessageId { get; set; }
-
-            public string EndpointNameOfReceivingEndpoint { get; set; }
-
             public string ConversationId { get; set; }
         }
     }
