+++ src/ServiceControl.AcceptanceTests/MultiInstance/When_message_searched_by_endpoint_by_messagetype.cs
@@ -1,6 +1,7 @@
 ﻿namespace ServiceBus.Management.AcceptanceTests.MultiInstance
 {
     using System.Collections.Generic;
+    using System.Linq;
     using System.Threading.Tasks;
     using EndpointTemplates;
     using Infrastructure.Settings;
@@ -8,29 +9,46 @@
     using NServiceBus.AcceptanceTesting;
     using NServiceBus.AcceptanceTesting.Customization;
     using NServiceBus.AcceptanceTests;
+    using NServiceBus.Settings;
     using NUnit.Framework;
     using ServiceControl.CompositeViews.Messages;
+    using ServiceControl.Infrastructure.Settings;
+    using Conventions = NServiceBus.AcceptanceTesting.Customization.Conventions;
 
-    class When_processed_message_multi_instance_searched_by_conversationId : AcceptanceTest
+    class When_message_searched_by_endpoint_by_messagetype : AcceptanceTest
     {
         [Test]
         public async Task Should_be_found()
         {
             SetInstanceSettings = ConfigureRemoteInstanceForMasterAsWellAsAuditAndErrorQueues;
 
-            await Define<MyContext>(Remote1, Master)
-                .WithEndpoint<Sender>(b => b.When((bus, c) => bus.SendLocal(new TriggeringMessage())))
+            var response = new List<MessagesView>();
+
+            var endpointName = Conventions.EndpointNamingConvention(typeof(ReceiverRemote));
+
+            //search for the message type
+            var searchString = typeof(MyMessage).Name;
+
+            var context = await Define<MyContext>(Remote1, Master)
+                .WithEndpoint<Sender>(b => b.When((bus, c) => bus.Send(new MyMessage())))
                 .WithEndpoint<ReceiverRemote>()
                 .Done(async c =>
                 {
-                    var result = await this.TryGetMany<MessagesView>($"/api/conversations/{c.ConversationId}", instanceName: Master);
-                    List<MessagesView> response = result;
-                    return c.ConversationId != null && result && response.Count == 2;
+                    var result = await this.TryGetMany<MessagesView>($"/api/endpoints/{endpointName}/messages/search/" + searchString, instanceName: Master);
+                    response = result;
+                    return result && response.Count == 1;
                 })
                 .Run();
+
+            var expectedRemote1InstanceId = InstanceIdGenerator.FromApiUrl(SettingsPerInstance[Remote1].ApiUrl);
+
+            var remote1Message = response.SingleOrDefault(msg => msg.MessageId == context.Remote1MessageId);
+
+            Assert.NotNull(remote1Message, "Remote1 message not found");
+            Assert.AreEqual(expectedRemote1InstanceId, remote1Message.InstanceId, "Remote1 instance id mismatch");
         }
 
-        private void ConfigureRemoteInstanceForMasterAsWellAsAuditAndErrorQueues(string instanceName, Settings settings)
+        void ConfigureRemoteInstanceForMasterAsWellAsAuditAndErrorQueues(string instanceName, Settings settings)
         {
             switch (instanceName)
             {
@@ -54,13 +72,13 @@
             }
         }
 
-        private string addressOfRemote;
-        private const string Master = "master";
-        private const string Remote1 = "remote1";
-        private static string AuditMaster = $"{Master}.audit";
-        private static string ErrorMaster = $"{Master}.error";
-        private static string AuditRemote = $"{Remote1}.audit";
-        private static string ErrorRemote = $"{Remote1}.error";
+        string addressOfRemote;
+        const string Master = "master";
+        const string Remote1 = "remote1";
+        static string AuditMaster = $"{Master}.audit";
+        static string ErrorMaster = $"{Master}.error";
+        static string AuditRemote = $"{Remote1}.audit";
+        static string ErrorRemote = $"{Remote1}.error";
 
         public class Sender : EndpointConfigurationBuilder
         {
@@ -70,22 +88,10 @@
                 {
                     c.AuditProcessedMessagesTo(AuditMaster);
                     c.SendFailedMessagesTo(ErrorMaster);
-
                     c.ConfigureTransport()
                         .Routing()
-                        .RouteToEndpoint(typeof(TriggeredMessage), typeof(ReceiverRemote));
+                        .RouteToEndpoint(typeof(MyMessage), typeof(ReceiverRemote));
                 });
-            }
-
-            public class TriggeringMessageHandler : IHandleMessages<TriggeringMessage>
-            {
-                public MyContext Context { get; set; }
-
-                public Task Handle(TriggeringMessage message, IMessageHandlerContext context)
-                {
-                    Context.ConversationId = context.MessageHeaders[Headers.ConversationId];
-                    return context.Send(new TriggeredMessage());
-                }
             }
         }
 
@@ -100,29 +106,31 @@
                 });
             }
 
-            public class TriggeredMessageHandler : IHandleMessages<TriggeredMessage>
+            public class MyMessageHandler : IHandleMessages<MyMessage>
             {
                 public MyContext Context { get; set; }
 
-                public Task Handle(TriggeredMessage message, IMessageHandlerContext context)
+                public ReadOnlySettings Settings { get; set; }
+
+                public Task Handle(MyMessage message, IMessageHandlerContext context)
                 {
-                    Context.ConversationId = context.MessageHeaders[Headers.ConversationId];
+                    Context.EndpointNameOfReceivingEndpoint = Settings.EndpointName();
+                    Context.Remote1MessageId = context.MessageId;
                     return Task.FromResult(0);
                 }
             }
         }
 
-        public class TriggeringMessage : ICommand
-        {
-        }
 
-        public class TriggeredMessage : ICommand
+        public class MyMessage : ICommand
         {
         }
 
         public class MyContext : ScenarioContext
         {
-            public string ConversationId { get; set; }
+            public string Remote1MessageId { get; set; }
+
+            public string EndpointNameOfReceivingEndpoint { get; set; }
         }
     }
 }
