+++ src/ServiceControl.AcceptanceTests/MultiInstance/When_message_searched_by_messageid.cs
@@ -1,7 +1,5 @@
 ﻿namespace ServiceBus.Management.AcceptanceTests.MultiInstance
 {
-    using System.Collections.Generic;
-    using System.Linq;
     using System.Threading.Tasks;
     using EndpointTemplates;
     using Infrastructure.Settings;
@@ -12,49 +10,22 @@
     using NServiceBus.Settings;
     using NUnit.Framework;
     using ServiceControl.CompositeViews.Messages;
-    using ServiceControl.Infrastructure.Settings;
 
-    class When_processed_message_multi_instance_by_messages : AcceptanceTest
+    class When_message_searched_by_messageid : AcceptanceTest
     {
         [Test]
         public async Task Should_be_found()
         {
             SetInstanceSettings = ConfigureRemoteInstanceForMasterAsWellAsAuditAndErrorQueues;
 
-            var response = new List<MessagesView>();
-
-            var context = await Define<MyContext>(Remote1, Master)
-                .WithEndpoint<Sender>(b => b.When(async (bus, c) =>
-                {
-                    await bus.Send(new MyMessage());
-                    await bus.SendLocal(new MyMessage());
-                }))
+            await Define<MyContext>(Remote1, Master)
+                .WithEndpoint<Sender>(b => b.When((bus, c) => bus.Send(new MyMessage())))
                 .WithEndpoint<ReceiverRemote>()
-                .Done(async c =>
-                {
-                    var result = await this.TryGetMany<MessagesView>("/api/messages/", instanceName: Master);
-                    response = result;
-                    return result && response.Count == 2;
-                })
+                .Done(async c => c.Remote1MessageId != null && await this.TryGetMany<MessagesView>("/api/messages/search/" + c.Remote1MessageId, instanceName: Master))
                 .Run();
-
-            var expectedMasterInstanceId = InstanceIdGenerator.FromApiUrl(SettingsPerInstance[Master].ApiUrl);
-            var expectedRemote1InstanceId = InstanceIdGenerator.FromApiUrl(SettingsPerInstance[Remote1].ApiUrl);
-
-            Assert.AreNotEqual(expectedMasterInstanceId, expectedRemote1InstanceId);
-
-            var masterMessage = response.SingleOrDefault(msg => msg.MessageId == context.MasterMessageId);
-
-            Assert.NotNull(masterMessage, "Master message not found");
-            Assert.AreEqual(expectedMasterInstanceId, masterMessage.InstanceId, "Master instance id mismatch");
-
-            var remote1Message = response.SingleOrDefault(msg => msg.MessageId == context.Remote1MessageId);
-
-            Assert.NotNull(remote1Message, "Remote1 message not found");
-            Assert.AreEqual(expectedRemote1InstanceId, remote1Message.InstanceId, "Remote1 instance id mismatch");
         }
 
-        void ConfigureRemoteInstanceForMasterAsWellAsAuditAndErrorQueues(string instanceName, Settings settings)
+        private void ConfigureRemoteInstanceForMasterAsWellAsAuditAndErrorQueues(string instanceName, Settings settings)
         {
             switch (instanceName)
             {
@@ -78,13 +49,13 @@
             }
         }
 
-        string addressOfRemote;
-        const string Master = "master";
-        const string Remote1 = "remote1";
-        static string AuditMaster = $"{Master}.audit";
-        static string ErrorMaster = $"{Master}.error";
-        static string AuditRemote = $"{Remote1}.audit";
-        static string ErrorRemote = $"{Remote1}.error";
+        private string addressOfRemote;
+        private const string Master = "master";
+        private const string Remote1 = "remote1";
+        private static string AuditMaster = $"{Master}.audit";
+        private static string ErrorMaster = $"{Master}.error";
+        private static string AuditRemote = $"{Remote1}.audit";
+        private static string ErrorRemote = $"{Remote1}.error";
 
         public class Sender : EndpointConfigurationBuilder
         {
@@ -94,24 +65,11 @@
                 {
                     c.AuditProcessedMessagesTo(AuditMaster);
                     c.SendFailedMessagesTo(ErrorMaster);
+
                     c.ConfigureTransport()
                         .Routing()
                         .RouteToEndpoint(typeof(MyMessage), typeof(ReceiverRemote));
                 });
-            }
-
-            public class MyMessageHandler : IHandleMessages<MyMessage>
-            {
-                public MyContext Context { get; set; }
-
-                public ReadOnlySettings Settings { get; set; }
-
-                public Task Handle(MyMessage message, IMessageHandlerContext context)
-                {
-                    Context.EndpointNameOfReceivingEndpoint = Settings.EndpointName();
-                    Context.MasterMessageId = context.MessageId;
-                    return Task.FromResult(0);
-                }
             }
         }
 
@@ -148,7 +106,6 @@
 
         public class MyContext : ScenarioContext
         {
-            public string MasterMessageId { get; set; }
             public string Remote1MessageId { get; set; }
 
             public string EndpointNameOfReceivingEndpoint { get; set; }
