+++ src/ServiceControl.AcceptanceTests/MultiInstance/When_message_searched_by_messagetype.cs
@@ -14,7 +14,7 @@
     using ServiceControl.CompositeViews.Messages;
     using ServiceControl.Infrastructure.Settings;
 
-    class When_processed_message_multi_instance_searched_by_messagetype : AcceptanceTest
+    class When_message_searched_by_messagetype : AcceptanceTest
     {
         [Test]
         public async Task Should_be_found()
